[devicelock] Keep watching the settings file if it is replaced. Fixes JB#36962
The standard practice for atomic writes is to write to a different
file and then rename it to replace the old.  This leads to
QFileSystemWatcher notifying the first time the file changes
and never again because the file it was watching is gone.  So
use inotify directly instead as that can tell us if the file
has been removed instead of merely changed and then we can watch
the new file instead.